separacion de carpetas para la pagina de restricciones
