+++ Sistema_Basico/src/App.tsx
@@ -5,7 +5,7 @@ import { ProfesoresPage } from "./components/profesores/ProfesoresPage";
 import { SalasPage } from "./components/salas/SalasPage";
 import { AsignaturasPage } from "./components/asignaturas/AsignaturasPage";
 import { HorariosPage } from "./components/horarios/HorariosPage";
-import { RestriccionesPage } from "./components/restricciones/RestriccionesPage";
+import { RestriccionesPage } from "./pages/RestriccionesPage/RestriccionesPage original";
 import { CursosPage } from "./components/cursos/CursosPage";
 
 export default function App() {
